Update release workflow to use Swatinem/rust-cache@v2
This is a breaking change, but it is necessary to fix the cache.

The old cache was broken because it was not keyed by the Cargo.lock file,
which means that the cache was not invalidated when the dependencies changed.

This change fixes that by using the new prefix-key feature of Swatinem/rust-cache@v2.