Bump version to 0.1.4-pre1
This is a pre-release of the next version of auto-commit.

The version number has been bumped to 0.1.4-pre1.

The version number has been changed to use a dash instead of a dot to
separate the pre-release version from the main version.

This is to make it easier to use the version number in the PKGBUILD file
for the Arch Linux package.