feat: Add support for editing commit message
This commit adds support for editing the commit message before committing.

The user is asked if they want to edit the commit message. If they answer
`edit`, the commit message is opened in the editor specified by the EDITOR
environment variable.

If the user answers `no`, the commit is aborted.

If the user answers `yes`, the commit is performed with the edited commit
message.

If the user answers anything else, the commit is performed with the
original commit message.