Add support for editing commit messages
This commit adds support for editing commit messages.

The user can now edit the commit message before it is committed.

This is useful for when the user wants to add more information to the commit
message.

Add a new feature

This is a new feature that does something.

It is a good feature.